tetragon/windows: Port unit tests in grpc package into  Windows
This PR changs the unit tests in grpc package to also run and succeed on Windows.
Some exec based unit tests are excluded due to ktime not being present in Windows exec event.
