+++ pkg/grpc/process_manager_test.go
@@ -5,8 +5,6 @@ package grpc
 
 import (
 	"context"
-	"encoding/base64"
-	"os"
 	"sync"
 	"testing"
 	"time"
@@ -17,7 +15,6 @@ import (
 	"github.com/cilium/tetragon/pkg/grpc/exec"
 	"github.com/cilium/tetragon/pkg/option"
 	"github.com/cilium/tetragon/pkg/process"
-	"github.com/cilium/tetragon/pkg/reader/node"
 	"github.com/cilium/tetragon/pkg/rthooks"
 	"github.com/cilium/tetragon/pkg/watcher"
 	"github.com/stretchr/testify/assert"
@@ -210,18 +207,4 @@ func TestProcessManager_GetProcessExec(t *testing.T) {
 			Setuid: val10000, Setgid: val10000,
 		},
 		exec.GetProcessExec(pi, false).Process.BinaryProperties)
-}
-
-func TestProcessManager_GetProcessID(t *testing.T) {
-	assert.NoError(t, os.Setenv("NODE_NAME", "my-node"))
-	node.SetExportNodeName()
-
-	err := process.InitCache(watcher.NewFakeK8sWatcher([]interface{}{}), 10, defaults.DefaultProcessCacheGCInterval)
-	assert.NoError(t, err)
-	defer process.FreeCache()
-	id := process.GetProcessID(1, 2)
-	decoded, err := base64.StdEncoding.DecodeString(id)
-	assert.NoError(t, err)
-	assert.Equal(t, "my-node:2:1", string(decoded))
-	assert.NoError(t, os.Unsetenv("NODE_NAME"))
 }
